- Created new hierarchy for analysis classes;
