+++ mip/src/main/java/br/edu/utfpr/cp/emater/midmipsystem/service/analysis/MIPSampleDefoliationAnalysisService.java
@@ -1,0 +1,5 @@
+package br.edu.utfpr.cp.emater.midmipsystem.service.analysis;
+
+public class MIPSampleDefoliationAnalysisService {
+    
+}
